requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.11.6
pydantic-core==2.33.2
pydantic-settings==2.1.0
boto3==1.29.3
mangum==0.19.0
email-validator==2.1.0.post1
greenlet==3.0.3
redis==5.0.1
aioapns==2.1
jinja2==3.1.2
cryptography==45.0.4
uvicorn==0.27.1
pytest==7.4.3
httpx==0.24.1
aiosqlite==0.19.0
pytest-cov==4.1.0